Added frame widget
This widget will eventually enable us to write movable "frames"
within a window.

I also noticed that many of the child bearing widgets in the Widget.ml
have similar handle code. So, i refactored most of it out into
a function called generic_passthough_handler. This handler is
appropriate for widgets that do not respond to events themselves,
but forward them to their children.